feat: consistent error api
