+++ src/actor/Claimer.ts
@@ -3,7 +3,7 @@ import {
   ERROR_MESSAGE_TYPE,
   ERROR_PE_MISMATCH,
   ERROR_PE_CREDENTIAL_MISSING,
-} from '../errorhandling/ObjectErrors'
+} from '../errorhandling/SDKErrors'
 import IPublicIdentity from '../types/PublicIdentity'
 import RequestForAttestation from '../requestforattestation/RequestForAttestation'
 import IClaim from '../types/Claim'
@@ -67,7 +67,7 @@ export async function createPresentation(
 
   // If privacy enhancement was required, but is not allowed, we can't create a presentation.
   if (!message.body.content.allowPE && requirePE) {
-    throw ERROR_PE_MISMATCH
+    throw ERROR_PE_MISMATCH()
   }
   const request = message.body
 
@@ -75,7 +75,7 @@ export async function createPresentation(
   if (request.content.allowPE) {
     const peCreds = credentials.map(c => c.privacyCredential)
     if (!noNulls(peCreds)) {
-      throw ERROR_PE_CREDENTIAL_MISSING
+      throw ERROR_PE_CREDENTIAL_MISSING()
     }
     const gabiPresentation = await identity.claimer.buildCombinedPresentation({
       credentials: peCreds,
