+++ src/claim/Claim.spec.ts
@@ -9,7 +9,7 @@ import {
   ERROR_CTYPE_HASH_NOT_PROVIDED,
   ERROR_HASH_MALFORMED,
   ERROR_ADDRESS_INVALID,
-} from '../errorhandling/ObjectErrors'
+} from '../errorhandling/SDKErrors'
 
 describe('Claim', () => {
   let identityAlice: Identity
@@ -119,7 +119,7 @@ describe('Claim', () => {
     expect(() => ClaimUtils.errorCheck(everything)).not.toThrow()
 
     expect(() => ClaimUtils.errorCheck(noCTypeHash)).toThrowError(
-      ERROR_CTYPE_HASH_NOT_PROVIDED
+      ERROR_CTYPE_HASH_NOT_PROVIDED()
     )
 
     expect(() => ClaimUtils.errorCheck(malformedCTypeHash)).toThrowError(
