+++ src/ctype/CType.utils.spec.ts
@@ -5,7 +5,7 @@ import {
   verifySchemaWithErrors,
 } from './CType.utils'
 import ICType from '../types/CType'
-import { ERROR_OBJECT_MALFORMED } from '../errorhandling/ObjectErrors'
+import { ERROR_OBJECT_MALFORMED } from '../errorhandling/SDKErrors'
 
 const ctypeInput = ({
   $id: 'http://example.com/ctype-1',
@@ -55,7 +55,7 @@ describe('CTypeUtils', () => {
     expect(verifySchemaWithErrors(badClaim, CTypeWrapperModel, [])).toBeFalsy()
     expect(() => {
       verifyClaimStructure(badClaim, ctypeInput)
-    }).toThrow(ERROR_OBJECT_MALFORMED)
+    }).toThrow(ERROR_OBJECT_MALFORMED())
   })
   it('verifies ctypes', () => {
     expect(verifySchema(ctypeWrapperModel, CTypeModel)).toBeTruthy()
