+++ src/ctype/CTypeMetadata.ts
@@ -7,7 +7,7 @@
 import ICTypeMetadata from '../types/CTypeMetadata'
 import { MetadataModel } from './CTypeSchema'
 import CTypeUtils from './CType.utils'
-import { ERROR_OBJECT_MALFORMED } from '../errorhandling/ObjectErrors'
+import { ERROR_OBJECT_MALFORMED } from '../errorhandling/SDKErrors'
 
 export default class CTypeMetadata implements ICTypeMetadata {
   public ctypeHash: ICTypeMetadata['ctypeHash']
@@ -23,7 +23,7 @@ export default class CTypeMetadata implements ICTypeMetadata {
    */
   public constructor(metadata: ICTypeMetadata) {
     if (!CTypeUtils.verifySchema(metadata, MetadataModel)) {
-      throw ERROR_OBJECT_MALFORMED
+      throw ERROR_OBJECT_MALFORMED()
     }
     this.metadata = metadata.metadata
     this.ctypeHash = metadata.ctypeHash
