+++ src/requestforattestation/RequestForAttestation.ts
@@ -39,7 +39,7 @@ import IRequestForAttestation, {
 import { IDelegationBaseNode } from '../types/Delegation'
 import IClaim from '../types/Claim'
 import IAttestedClaim from '../types/AttestedClaim'
-import * as ObjectErrors from '../errorhandling/ObjectErrors'
+import * as SDKErrors from '../errorhandling/SDKErrors'
 
 function hashNonceValue(
   nonce: string,
@@ -137,7 +137,7 @@ export default class RequestForAttestation implements IRequestForAttestation {
     session: ClaimerAttestationSession | null
   }> {
     if (claim.owner !== identity.getAddress()) {
-      throw ObjectErrors.ERROR_IDENTITY_MISMATCH
+      throw SDKErrors.ERROR_IDENTITY_MISMATCH()
     }
 
     let peRequest: AttestationRequest | null = null
@@ -279,7 +279,7 @@ export default class RequestForAttestation implements IRequestForAttestation {
   public removeClaimProperties(properties: string[]): void {
     properties.forEach(key => {
       if (!this.claimHashTree[key]) {
-        throw ObjectErrors.ERROR_CLAIM_HASHTREE_MISMATCH(key)
+        throw SDKErrors.ERROR_CLAIM_HASHTREE_MISMATCH(key)
       }
       delete this.claim.contents[key]
       delete this.claimHashTree[key].nonce
@@ -327,7 +327,7 @@ export default class RequestForAttestation implements IRequestForAttestation {
     Object.keys(input.claim.contents).forEach(key => {
       const value = input.claim.contents[key]
       if (!input.claimHashTree[key]) {
-        throw ObjectErrors.ERROR_CLAIM_HASHTREE_MALFORMED
+        throw SDKErrors.ERROR_CLAIM_HASHTREE_MALFORMED()
       }
       const hashed: NonceHash = input.claimHashTree[key]
       validateNonceHash(hashed, value, `hash tree property ${key}`)
@@ -338,11 +338,11 @@ export default class RequestForAttestation implements IRequestForAttestation {
 
     // check claim hash
     if (!RequestForAttestation.verifyRootHash(input)) {
-      throw ObjectErrors.ERROR_ROOT_HASH_UNVERIFIABLE
+      throw SDKErrors.ERROR_ROOT_HASH_UNVERIFIABLE()
     }
     // check signature
     if (!RequestForAttestation.verifySignature(input)) {
-      throw ObjectErrors.ERROR_SIGNATURE_UNVERIFIABLE
+      throw SDKErrors.ERROR_SIGNATURE_UNVERIFIABLE()
     }
 
     return true
